Fix bugs inadvertently introduced in earlier work; Fixes DDR-1561.
The work done in commit 8ebf69b290a1f881f49ae87d8e583f7b80cb35a2 for DDR-1444 resulted
in the 'research_help_contact' field becoming an editable admin metadata field in all
contexts, not just for collections and in Metafile File Upload.  This caused the bug reported
in DDR-1561 as well as making that field editable in contexts where it should not have been.